Feature: creating function structure for fetching
